check if file already exists before re-downloading

+++ start.sh
@@ -15,12 +15,14 @@ fi
 
 cd /workspace/text-generation-webui/models
 
-wget https://huggingface.co/TheBloke/guanaco-65B-GGML/resolve/main/guanaco-65B.ggmlv3.q8_0.z01
-wget https://huggingface.co/TheBloke/guanaco-65B-GGML/resolve/main/guanaco-65B.ggmlv3.q8_0.zip
-zip -FF guanaco-65B.ggmlv3.q8_0.zip --out guanaco-65B.ggmlv3.q8_0.full
-rm guanaco-65B.ggmlv3.q8_0.z*
-unzip guanaco-65B.ggmlv3.q8_0.full
-rm guanaco-65B.ggmlv3.q8_0.full
+if [ ! -f guanaco-65B.ggmlv3.q8_0.bin ]; then
+    wget https://huggingface.co/TheBloke/guanaco-65B-GGML/resolve/main/guanaco-65B.ggmlv3.q8_0.z01
+    wget https://huggingface.co/TheBloke/guanaco-65B-GGML/resolve/main/guanaco-65B.ggmlv3.q8_0.zip
+    zip -FF guanaco-65B.ggmlv3.q8_0.zip --out guanaco-65B.ggmlv3.q8_0.full
+    rm guanaco-65B.ggmlv3.q8_0.z*
+    unzip guanaco-65B.ggmlv3.q8_0.full
+    rm guanaco-65B.ggmlv3.q8_0.full
+fi
 
 
 if [[ $JUPYTER_PASSWORD ]]
